Moved all android tests into a *.main package so that it won't include screenshots package.
